Develop2

+++ Server/deployment/application_stop.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 echo "Stopping any existing node servers"
-pkill node
+if pgrep node; then pkill node; fi
